Convert UI text from Turkish to English
Replaces Turkish UI strings with English equivalents across multiple components, including chat, sidebar, navbar, modals, and dashboard. Also refactors ConversationMemory node configuration: removes the modal and introduces a new config form and supporting files under the ConversationMemory directory.
